Integrate Yuvi work for frontend Add permissions to all methods (partial done) Split categories into umbrellas, fields, links (models, controllers, views) (80% done) Create error pages for permisions Add access control on groups

+++ base/application/controllers/admin_panel/Brandleads.php
@@ -9,12 +9,12 @@ if (!defined('BASEPATH')) {
 }
 
 /**
- * File:    ~/application/models/admin_panel/Brandleads_model_admin.php
+ * File: ~/application/controller/admin_panel/Brandleads.php
  *
  * A controller for brandleads
  * 
  * @package      Skearch
- * @author       Iftikhar Ejaz <i.ejaz@skearch.net>
+ * @author       Iftikhar Ejaz <ejaziftikhar@gmail.com>
  * @copyright    Copyright (c) 2020
  * @version      2.0
  */
@@ -22,14 +22,19 @@ class Brandleads extends MY_Controller
 {
 
     /**
-     * Undocumented function
+     * Constructor
      */
     public function __construct()
     {
         parent::__construct();
 
-        if (!$this->ion_auth->is_admin()) {
-            // redirect them to the login page
+        if (!$this->ion_auth->logged_in()) {
+            // redirect to the admin login page
+            redirect('admin/auth/login');
+        }
+
+        if (!$this->ion_auth->in_group($this->config->item('staff', 'ion_auth'))) {
+            // redirect to the admin login page
             redirect('admin/auth/login');
         }
 
@@ -38,15 +43,21 @@ class Brandleads extends MY_Controller
 
     public function index()
     {
+        if (!$this->ion_auth_acl->has_permission('brandleads_get') && !$this->ion_auth->is_admin()) {
+            // set page title
+            $data['title'] = ucwords('access denied');
+            $this->load->view('admin_panel/errors/error_403', $data);
+        } else {
 
-        if (!file_exists(APPPATH . '/views/admin_panel/pages/brands/brandleads.php')) {
-            show_404();
+            if (!file_exists(APPPATH . '/views/admin_panel/pages/brands/brandleads.php')) {
+                show_404();
+            }
+
+            $data['title'] = ucfirst("Brand Leads");
+
+            // Load page content
+            $this->load->view('admin_panel/pages/brands/brandleads', $data);
         }
-
-        $data['title'] = ucfirst("Brand Leads");
-
-        // Load page content
-        $this->load->view('admin_panel/pages/brands/brandleads', $data);
     }
 
     /**
@@ -56,19 +67,21 @@ class Brandleads extends MY_Controller
      */
     public function get()
     {
-        $brandleads = $this->Brandleads->get();
-        $total_brands = sizeof($brandleads);
-        $result = array(
-            'iTotalRecords' => $total_brands,
-            'iTotalDisplayRecords' => $total_brands,
-            'sEcho' => 0,
-            'sColumns' => "",
-            'aaData' => $brandleads,
-        );
+        if ($this->ion_auth_acl->has_permission('brandleads_get') or $this->ion_auth->is_admin()) {
+            $brandleads = $this->Brandleads->get();
+            $total_brands = sizeof($brandleads);
+            $result = array(
+                'iTotalRecords' => $total_brands,
+                'iTotalDisplayRecords' => $total_brands,
+                'sEcho' => 0,
+                'sColumns' => "",
+                'aaData' => $brandleads,
+            );
 
-        $this->output
-            ->set_content_type('application/json')
-            ->set_output(json_encode($result));
+            $this->output
+                ->set_content_type('application/json')
+                ->set_output(json_encode($result));
+        }
     }
 
     /**
@@ -79,6 +92,10 @@ class Brandleads extends MY_Controller
      */
     public function delete($id)
     {
-        $this->Brandleads->delete($id);
+        if (!$this->ion_auth_acl->has_permission('brandleads_delete') && !$this->ion_auth->is_admin()) {
+            echo json_encode(-1);
+        } else {
+            $this->Brandleads->delete($id);
+        }
     }
 }
